+++ packages/terraform/src/executors/fmt/compat.ts
@@ -1,0 +1,5 @@
+import { convertNxExecutor } from '@nx/devkit'
+
+import fmtExecutor from './fmt.impl'
+
+export default convertNxExecutor(fmtExecutor)