+++ packages/gcp-cloud-run/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [6.1.2](https://github.com/TriPSs/nx-extend/compare/gcp-cloud-run@6.1.1...gcp-cloud-run@6.1.2) (2023-08-31)
+
+
+
+## [6.1.1](https://github.com/TriPSs/nx-extend/compare/gcp-cloud-run@6.1.0...gcp-cloud-run@6.1.1) (2023-08-25)
+
+
+
 # [6.1.0](https://github.com/TriPSs/nx-extend/compare/gcp-cloud-run@6.0.0...gcp-cloud-run@6.1.0) (2023-06-22)