+++ packages/gcp-deployment-manager/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [5.1.2](https://github.com/TriPSs/nx-extend/compare/gcp-deployment-manager@5.1.1...gcp-deployment-manager@5.1.2) (2023-08-31)
+
+
+
+## [5.1.1](https://github.com/TriPSs/nx-extend/compare/gcp-deployment-manager@5.1.0...gcp-deployment-manager@5.1.1) (2023-08-25)
+
+
+
 # [5.1.0](https://github.com/TriPSs/nx-extend/compare/gcp-deployment-manager@5.0.0...gcp-deployment-manager@5.1.0) (2023-06-22)