+++ packages/gcp-functions/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [9.3.6](https://github.com/TriPSs/nx-extend/compare/gcp-functions@9.3.5...gcp-functions@9.3.6) (2023-09-14)
+
+
+
+## [9.3.5](https://github.com/TriPSs/nx-extend/compare/gcp-functions@9.3.4...gcp-functions@9.3.5) (2023-09-07)
+
+
+
+## [9.3.4](https://github.com/TriPSs/nx-extend/compare/gcp-functions@9.3.3...gcp-functions@9.3.4) (2023-08-31)
+
+
+
+## [9.3.3](https://github.com/TriPSs/nx-extend/compare/gcp-functions@9.3.2...gcp-functions@9.3.3) (2023-08-25)
+
+
+
 ## [9.3.2](https://github.com/TriPSs/nx-extend/compare/gcp-functions@9.3.1...gcp-functions@9.3.2) (2023-08-06)