+++ packages/github-pages/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [4.1.2](https://github.com/TriPSs/nx-extend/compare/github-pages@4.1.1...github-pages@4.1.2) (2023-08-31)
+
+
+
+## [4.1.1](https://github.com/TriPSs/nx-extend/compare/github-pages@4.1.0...github-pages@4.1.1) (2023-08-25)
+
+
+
 # [4.1.0](https://github.com/TriPSs/nx-extend/compare/github-pages@4.0.0...github-pages@4.1.0) (2023-06-22)