+++ packages/strapi/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [6.1.7](https://github.com/TriPSs/nx-extend/compare/strapi@6.1.6...strapi@6.1.7) (2023-09-14)
+
+
+
+## [6.1.6](https://github.com/TriPSs/nx-extend/compare/strapi@6.1.5...strapi@6.1.6) (2023-09-07)
+
+
+
+## [6.1.5](https://github.com/TriPSs/nx-extend/compare/strapi@6.1.4...strapi@6.1.5) (2023-08-31)
+
+
+
+## [6.1.4](https://github.com/TriPSs/nx-extend/compare/strapi@6.1.3...strapi@6.1.4) (2023-08-25)
+
+
+
 ## [6.1.3](https://github.com/TriPSs/nx-extend/compare/strapi@6.1.2...strapi@6.1.3) (2023-08-06)