+++ packages/translations/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [10.1.3](https://github.com/TriPSs/nx-extend/compare/translations@10.1.2...translations@10.1.3) (2023-09-14)
+
+
+
+## [10.1.2](https://github.com/TriPSs/nx-extend/compare/translations@10.1.1...translations@10.1.2) (2023-08-31)
+
+
+
+## [10.1.1](https://github.com/TriPSs/nx-extend/compare/translations@10.1.0...translations@10.1.1) (2023-08-25)
+
+
+
 # [10.1.0](https://github.com/TriPSs/nx-extend/compare/translations@10.0.0...translations@10.1.0) (2023-06-22)